only download if not exist

+++ download_opencv_native.sh
@@ -1,9 +1,13 @@
 DIR=`pwd`
-DESTINATION=$DIR/android/src/main
-curl -O https://jaist.dl.sourceforge.net/project/opencvlibrary/3.4.4/opencv-3.4.4-android-sdk.zip
+DESTINATION=$DIR/android/src/main/jniLibs
 
-unzip opencv-3.4.4-android-sdk.zip
-mkdir $DIR/android/src/main/jniLibs
-cp -r $DIR/OpenCV-android-sdk/sdk/native/libs/ $DESTINATION/jniLibs
-rm -rf opencv-3.4.4-android-sdk.zip
-rm -rf OpenCV-android-sdk/
+if [ ! -d "$DESTINATION" ]; then
+  # Only download if destination does not exist
+  curl -O https://jaist.dl.sourceforge.net/project/opencvlibrary/3.4.4/opencv-3.4.4-android-sdk.zip
+
+  unzip opencv-3.4.4-android-sdk.zip
+  mkdir $DESTINATION
+  cp -r $DIR/OpenCV-android-sdk/sdk/native/libs/ $DESTINATION
+  rm -rf opencv-3.4.4-android-sdk.zip
+  rm -rf OpenCV-android-sdk/
+fi
